Add base for application callbacks
In order to get a better control on the end application, we need to have
a place to put callbacks. This commit adds a base for this.

The commit itself does nothing: it creates the base for the callbacks,
but they are never setup or called. This will be done in a future
commit.

Note that the added `modbus_t *ctx` parameter is not used by this
default implementation. The intention for it is to be filled by the
application implementation, so that the callbacks can access the context
for whichever purpose they might need.

Signed-off-by: Esa Laakso <esa.laakso@fidelix.com>