Use defined register callbacks
A previous commit added default callbacks for the registers. This commit
takes them into use.

The default callbacks are defined in the _modbus_init_default_callbacks
function, and we always expect it to have been called. The code does no
checks past _modbus_init_common if they are defined or not. This is to
keep the code simple and fast.

Signed-off-by: Esa Laakso <esa.laakso@fidelix.com>