Change the code structure of mapping allocation
In order to

* Allow handling custom data storage
* Make the slightly more efficient to be read

add a new function `modbus_mapping_new_address_only` which only sets the
registry boundary and the number of registers. This will be used to
create the base for modbus_mapping_t. It will

* Allocate the base of the struct
* Sets the register boundary (start address and number of registers)
* Sets the mapping to default NULL

There are two cases where this is used. The first is the current code
base uses the function to set the base information. It will then proceed
to allocate the registers. The second case is when the user wants to use
purely callbacks to handle the data. In that case, the user will call
this function and then manage mapping or storage in their
implementation.

Signed-off-by: Esa Laakso <esa.laakso@fidelix.com>